Remove Loader and Tweets List Fragment
