--- twitter-clone/app/src/main/java/io/github/vnicius/twitterclone/ui/common/fragments/LoaderFragment.kt
@@ -1,28 +1,0 @@
-package io.github.vnicius.twitterclone.ui.common.fragments
-
-import android.os.Bundle
-import android.support.v4.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import io.github.vnicius.twitterclone.R
-
-/**
- * [Fragment] to show the loader animation
- */
-class LoaderFragment : Fragment() {
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_loader, container, false)
-    }
-
-    companion object {
-        @JvmStatic
-        fun newInstance() =
-            LoaderFragment()
-    }
-}
