+++ twitter-clone/app/src/main/res/layout/partial_profile_content.xml
@@ -132,12 +132,18 @@
 
         </LinearLayout>
 
-        <FrameLayout
-                android:id="@+id/fl_profile_content_fragment_layout"
+        <android.support.v7.widget.RecyclerView
+                android:id="@+id/rv_profile_content_tweets_list"
                 android:layout_width="match_parent"
-                android:layout_height="wrap_content">
+                android:layout_height="wrap_content"
+                android:visibility="gone"
+                tools:listitem="@layout/item_tweet" />
 
-        </FrameLayout>
+        <include
+                android:id="@+id/inc_profile_content_tweets_spinner"
+                layout="@layout/fragment_loader"
+                android:visibility="gone" />
+
     </LinearLayout>
 
 </android.support.v4.widget.NestedScrollView>
